requirements.txt
Pillow>=9.0.0
numpy>=1.21.5
skimage>=0.19.2
Shapely>=1.8.0
torch==1.9.0
torchvision==0.10.0
opencv-python>=4.6.0
pandas>=1.3.5
strawberry-graphql>=0.165.1
fastapi>=0.95.0
prisma>=0.8.2
matplotlib>=3.2.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
tqdm>=4.64.0
psutil
